IA-1520: refactor SpeedDialInstance
- isolate actions in hooks
- split up helpers and network code in separate files
