+++ hat/assets/js/apps/Iaso/types/utils.ts
@@ -11,3 +11,4 @@ export type GenericObject = Record<string, any>;
 export type Nullable<T> = T | null;
 // This type might be useful when typing eg useState
 export type Optional<T> = T | undefined;
+export type ClassNames = Record<string, string>;
